Explain default domain in docs for other services
The documentation for other services previously made an assertion that
(paraphrasing) "using v3 tokens on v2 works fine." That statement comes
with a laundry list of caveats, starting with the "default" domain.

This patch revises the documentation to introduce the default domain
first, and then enumerates the caveats that come as a result of mixing
and matching v2 and v3.

There's also some general RST syntax improvements, grammar improvements,
and line wrapping politely at 79 characters like your mother taught you
to do.

Change-Id: I342092b9c9971d44438fa83f547d1f5ddd204ced