Handle fernet payload timestamp differences
The fernet payload tests assumed that the timestamp string that
was put into the payload would be exactly the same as the timestamp
string resulting from parsing the payload, but this isn't the case
since the timestamp string goes through a conversion to float and
back and a tiny amount of precision may be lost.

The tests are changed to allow a tiny difference in timestamp
strings.

This was causing python3.4 unit tests to fail sometimes.

bp python3

Change-Id: Icc94c7a9b322da6b9e098f111e9e254f10cc23bf