Move oauth1 sql migrations to common
This is the most basic way to move the oauth1 migrations into
the keystone core migrations path. Check to see if the migration was
performed already via the extension and if so skip performing it again.

DocImpact: You should no longer run the migrations for the oauth extension

Change-Id: I8cea6748e7a5ae2b01320b4d6a9866e60ff2ea74
Implements: bp move-extensions