Add caching to get_catalog
This commit adds caching to the get_catalog call, per user ID and
project ID. Both the v2 and v3 catalogs are built based on the result of
this method, so this provides caching for both APIs.

As a result, this should provide a performance benefit to Fernet-based
deployments, which must otherwise rebuild the catalog on each token
validation.

Tested on top of the current stable/liberty, this patch improves the
time of cacheable Fernet token creation requests by 16.7% and of
Fernet token validation requests by 7.5%:

  https://gist.github.com/dolph/3bf24039b83a147eeb5c

Change-Id: I147c5b8e7dcfe1786fe8a534182dbeef33925eac
Co-Authored-By: Dolph Mathews <dolph.mathews@gmail.com>
Related-Bug: 1489061