Fix UnexpectedError exceptions to use debug_message_format
The point of UnexpectedError is to hide the real error from the
client so that extra information won't be sent to an attacker.
This only happens when the debug_message_format is set rather
than message_format. Several UnexpectedError implementations
were setting message_format rather than debug_message_format so
this has been corrected.

Change-Id: I77efba7437b56936041c26c35e9c68760806455e