More useful message when using direct driver import
In the Liberty release, we switched to using entrypoints for
specifying the driver or auth plugin. The deprecation message
didn't mention the driver name that was specified, nor did it
mention where to look for the expected names, so it was not
user friendly.

Closes-Bug: 1513102
Change-Id: I02e265684b26686523da9d648b37675feb052978
(cherry picked from commit c73907485a3d7c5319378d2e3a5e11593f66a930)