Use list_role_assignments to get projects/domains for user
Now that list_role_assignments is paramaterized, we can use it
to replace bespoke code that also tries to evaluate assignments
and inheritance, such as getting projects and domains for a user.

This not only makes assignment manager easier to maintain, but
will also simplify upcoming changes to inheritance.

Co-Authored-By: Samuel de Medeiros Queiroz <samueldmq@gmail.com>

Partially Implements: blueprint assignment-manager-cleanup

Change-Id: Ie5441ab469668aa23418ca2290c43693e04142dc