Enhance manager list_role_assignments to support group listing
With the enahnced list_role_assignment manager method we have
already removed most of the duplication of inheritance logic when
listing roles. However, for federated operations there was a
class of query focused around assignments that derive from the
set of groups (i.e. those a federation user is given membership
of) that was not supported by list_role_assignments.

This patch adds that support, removes the duplicative
driver methods and re-implements the manager methods to call
list_role_assignments.

Implements: blueprint assignment-manager-cleanup
Change-Id: I9d76878963b178ec474b0c89be2ad2024bdd10d5