Create new version of assignment driver interface
In preparation for removing a series of driver methods in
the rationalization of assignment in follow-on patches, we
first create a V9 version of the driver interface. In this
patch, both V8 and V9 are identical.

To create the new V9 driver we:

1) Copy the sql V8 driver into its own directory to provide a
test for support of a legacy driver
2) Increase the version of the (new) standard driver to V9
3) Create a wrapper for V8 legacy drivers, so that the latest
manager can be driver version-less
4) Create tests that run the standard sql Identity tests
against the V8 driver to show, going forward, it is still working.
Since our existing test strategy assumes that we never change the
sqlalchemy models during a test run, this turned out to be slight
more complex than anticipated - so a new tox envlist target for legacy
drivers is added.

Partially Implements: blueprint assignment-manager-cleanup
Change-Id: Iefa13b64084b8e71c881ae08e1ec434d24756c17