Add testcases to check cache invalidation in endpoint filter extension
keystone now has the ability to cache catalog, and will invalidate
the cache when there is any change about the catalog, i.e. add endpoints
to project or remove endpoints from project. But the testcase was missed.

The patch focus on endpoint project association and leave endpoint_group
project association as it is. since these two APIs should impact the
endpoints in theory but it's doesnot in practice. So, add some testcases
to test its impact on catalog is impossible.

There is another bug filed [1] and testcases for endpoint group will be
added when that bug is fixed.

[1] https://bugs.launchpad.net/keystone/+bug/1516469

Change-Id: Ie85679111e0f8b9d00e11155594db391eb92c09f
Closes-Bug: #1516468