Fix test_crud_user_project_role_grants
That test, located at test_v3_assignment.py, has
become invalid, as per FIXME note attached to it.

The issue is that it was deleting the role
assignment the current user needed to to perform
the test itself.

This patch works around this by creating a second
role assignment and using it for testing.

Change-Id: I2c6a3b2e23d9ea026965350e44182bc0df80924e