Add checks for domain scoped data creep
Previously, the assertValidDomainScopedTokenResponse method only ensured
specific attributes were in the token response. These checks didn't ensure
that the token scope never grew.

This change makes it so that the assertion will fail if extra attributes are
added to the token response. This should help us be more aware of changes that
have token response data creep by building the check into the tests.

Change-Id: I43c86837f465f813da0985033a5af9d15d76eddc
Partial-Bug: 1224273