Add checks for project scoped data creep to tests
Previously, the assertValidProjectScopedTokenResponse method only ensured
specific attributes were in the token response. These checks didn't ensure
that the token scope never grew.

This change makes it so that the assertion will fail if extra attributes are
added to the token response. This should help us be more aware of changes that
have token response data creep by building the check into the tests.

Change-Id: I4ebb86af973ed6af001373756a69af9586bdefcf
Closes-Bug: 1224273