Verify that attribute `enabled` equals True
In change I2b67c34796c54b0963fbc7020d4e78ed4f71c7f3 all occurences of
assertEqual(True, ...) were changed to assertTrue(...). Although it's
fine for most of the changes, one of them should have stayed as it is,
because assertTrue converts the passed argument to bool(). So, a call
assertTrue(-42) is the same as assertTrue(bool(-42)), and we want
AssertionError to be raised in this case.

Use assertIs instead of assertTrue were appropriate. Tell others not
to change it.

Change-Id: Ifec683e7b5b728244f0444aa24a90994b1ce5153