Parameter to return audit ids only in revocation list
The revocation list will only include audit IDs and not token
IDs if ?audit_id_only is on the request. Also, the data won't be
obfuscated since the audit_ids aren't useful for auth. If the clients
(typically auth_token) only request the revocation list with audit IDs
only then the deployer doesn't have to do pki_setup to use the
revocation list.

Closes-Bug: 1317302
Change-Id: I197df21098f545b27163c9456917491561b53abb