Create V9 version of federation driver interface
In preparation for api changes for filtering, a V9 version of
the federation driver interface is created.

This patch also stops trying to load any of the contrib sql
models in our test runs - since these were a duplicative load
of the same models, since all contrib sql backends have moved
to the core tree. Not loading the contrib models is a
requirement for being able to test the V8 federation interface,
since otherwise two different models get loaded.

In addition, this patch fixes up the setup of restful test cases
to allow legacy driver overrides to be used in such tests. This
is useful for components (such as federation) that do not have
good manager level testing due to their single backend driver.

Change-Id: I179c5b86e42b8ac67b630b6871f1819c6f62c96b
Partial-Bug: #1525317