Modify implied roles to honor domain specific roles
The logic for processing domain specific roles is the same as
regular implied roles, except for the fact that domain specifc
roles themselves should not be returned by the manager level
list_role_assignments() in effective mode, hence ensuring that
the won't be placed in the token.

This patch makes the above changes, and adds tests to ensure this.

A follow-on patch will update the policy rules for domain specific
role assignment as well as provide release notes.

Partially implements: blueprint domain-specific-roles

Change-Id: I8c5d0dfd329a84a0525d625dc3b0c2c9325ab6bb