Revert "Validate domain ownership for v2 tokens"
This reverts commit c4723550aa95be403ff591dd132c9024549eff10.

This revert is being proposed as it breaks behavior that real-world
deployments rely on. The deployments requested the V2 token with
user_id and tenantId and then used the V2 token for the
non-default-domain user to access swift.

While the deployment is being encouraged to fix their code to use V3,
this is behavior that was supported and used.

Conflicts (new file post Liberty):
    keystone/tests/unit/test_v3_resource.py

Closes-Bug: 1527759
Change-Id: I4a303a5fcc8c2dacef5960e9e26ad9402f34a790