Fix incorrect signature in federation legacy V8 wrapper
The abstract signature for the create_mapping driver method in the
Liberty V8 driver interface was wrong, which was then inadvertently
used to create the V9 wrapper. This fixes that issue and adds attitional
legacy tests to ensure the federation V8 CRUD operations are covered.

There are numerous other inconsistency in the docstrings of the
federation abstract methods, these are resolved in a follow on patch.

Closes-bug: #1533346

Change-Id: I1d0aa2a652c681512e5268567b994ce3459e8eac