Update v3policysample tests to use admin_project not special domain_id
We now support the special admin_project that can be used to grant
cloud-wide powers. The tests on the v3cloudsample had been half
updated to use this new facility, but in fact were still using the
old "patch a domain id in the policy file" approach. As well as
not testing the new functionality, the current tests were causing
problems elsewhere since they used a non-UUID domain_id.

Closes-Bug: #1533330
Change-Id: Ic116cf8715130f6ed6bd5380c1b31d5ef0ca154d