Use the driver to get limits
@response_truncated was used to set the limit of returned entries. It
asked the driver about the limit and set it to hints. With
domain-specific configs, there are multiple driver instances and each of
them carries domain-specific config.

However, with domain-specific configs, the driver is not yet configured
at that point, because sometimes the manager needs to perform additional
actions in order to understand what domain it works with. Because of
that, @response_truncated always got the limit from the default driver,
not from the one actually used for the domain.

Move the logic of setting the limit from the decorator to a private
method, call it after determining the domain and driver.

Change-Id: I1748d491b047e33712380da731c272f9d471ec0a
Closes-Bug: 1495669