Add linters environment, keep pep8 as alias
Rename the pep8 environment to linters.
Use linters to signify that this runs a variety of
lint like jobs (name is used by other repos already).
Add pep8 back in as alias, so that developers get the same experience
running either of the environments.
When users run pep8, an information about linters is given.

Remove {posargs} from flake8, it's not used and breaks the alias.

Change-Id: Ic9efdcb13978d2f29312a4a1f0fdd859d55c8398