Update mod_wsgi + cache config docs
I was asked if dogpile.cache.memcached was recommended for Apache httpd
with mod_wsgi because our docs did not make it clear that
dogpile.cache_memcached_pool wasn't necessary. This patch adds an
explicit recommendation.

Along the way, I also cleaned up a few RST formatting issues and
consistently referred to "Apache httpd with ``mod_wsgi``", which is how
Apache *usually* does it themselves (sometimes HTTPD is in all caps, but
it's never a literal like module names are).

Change-Id: Ic562853bacc13f70712e0e61f8aeb8dffc75ab2e