exception sensitive cache/audit changes
Several places were doing things between when a cached resource was
changed and when the cache was invalidated, which could lead to
problems, e.g. if the code in between the two raised an exception,
or if an audit event resulted in the resource being requeried before
the cache was invalidated. Review comments on the original patch also
pointed out that audit events would not be sent in some exeption cases.

This change does cache invalidation as quickly as possible after a
change and uses try/finally logic to ensure that audit events will
be sent even in exception cases.

It also removes a stale TODO and an unnecessary duplicate assignment
removal in the affected code.

Change-Id: If9a91da42818fe37c975f86be47a2e49da356e41