Revert "Unit test for checking cross-version migrations compatibility"
This reverts commit 8234b92e5751c22e39d38eb4628c54c9b86d13de.

Due to discussions at the midcycle and based upon the complexity of
cross-version schema support, it has been determined that this
simple test is unfortunately both insufficient and the incorrect
approach to ensure schema compatibility. We have discussed and
come to a decision on how we are going both prove out the
plan for cross-version support and the appropriate path for
upgrade. The base feature of what this test encompasses will
instead be migrated over to a full separate gate/check job
that will be able to handle the more complex tasks of
ensuring schema upgrades make sense. For now this
test is inaccurate and missing changes such as FK retargeting
and blocking the ability to perform work that is needed
for the Mitaka cycle. The plan made at the midcycle
will continue through Mitaka with the plan to enact
enforcement in Newton.

Change-Id: I9ae0caed991b18600b70d408f32d837c92d059c2