Convert assignment.root_role config option to list of strings
The assigment.root_role config option was previously a single string
which specified a role name that was prohibited from being added as an
implied role. By default it was 'admin'. For greater flexibility we
now permit a list of role names that are prohibited from being
implied.

Summary of changes:

* Change assigment.root_role from cfg.StrOpt to cfg.ListOpt.
  ListOpt is preferred over MultiStrOpt because of config file formatting.
  Update help for option.

* Change assigment.root_role name to assignment.prohibited_implied_role

* Change test for implied role name from string equality to
  membership in list of strings.

* Expand ImpliedRolesTests.test_root_role_as_implied_role_forbidden()
  unit test to test 2 prohibited implied role names and 1 valid
  implied. role name.

Change-Id: Idfe14080e2f1ec1e89b85d8f5f00aad187f1fd22
Closes-Bug: #1541540
Signed-off-by: John Dennis <jdennis@redhat.com>