db_sync doesn't create default domain
The reason db_sync needed to create the default domain is because we
needed a domain for existing v2 users. Since the migrations don't
add the domain_id to users anymore there's no need to create the
default domain.

DocImpact -- The install guide should be updated to say to use
 keystone-manage bootstrap or to create the default domain if the
 deployment is going to support v2.

Change-Id: I65860fe989ac2456b73bcc12fd02643564b24574