Add ability to send notifications for actors
Previously, keystone didn't have the ability to send notifications for
actor/target relationships due to the structure of the payload. The payload was
always expected to be the ID and type of one entity. In notifications that have
relationships (like adding or removing users from groups), it is useful to know
who was added to what.

This commit introduces the ability to send information about who was added or
removed from an entity.

Change-Id: Iea4437a265f1910a1171b9bcac99d853f5015440
Partial-Bug: 1552639