Split assignment backend tests
Where to put a new test ? Is this behavior covered yet ? These and
other questions require understanding the test coverage to be answered.

test_backend.py is a huge test file, making it hard to understand what
cases are currently covered for each backend.

This chain of changes proposes to split it by the backends it covers;
the tests will be placed into 'unit/{backend}/test_backends.py'.

This change splits the tests for the assignment backend.

Change-Id: I43ecea73d8c4ee70dd861d010282affb2ec3232a