Remove _disable_domain from the resource API
The _disable_domain method in was only there to act as a method we could put a
decorator on for notifications. The notification decorator has been replaced
with an inline statement for sending the notification. This means we no longer
need to have a specific method for the purpose of a decorator, so we can remove
_disable_domain.

Change-Id: I86ec5d7bc317454b2a95e1afc2b380e292acf363