Fixes a few LDAP tests to actually run
It looks like the original intent of overriding these test methods was
to actually make the test run. The direct parent, BaseLDAPIdentity, was
overriding them from its parent in order to show they raise an
exception. The modified test class tests the fix for this behavior and
want to execute the original test, not the overridden one. Unfortunately
the test method was never actually called.

Change-Id: Ib312e75f5a61ef7518177608acce56dfdc05afbf