Clarify virtualenv setup in developer docs
The Developer Environment documentation has an error where it explains
how to create a virtualenv with tox. If we follow instructions, we end
up with a nasty traceback from tox[1]. This is because the 'venv'
environment in tox.ini is expecting positional arguments and has no
default for when none are provided. This patch cribs from the Ironic
developer docs[2] to add a positional argument to make tox happy.

[1] http://paste.openstack.org/show/489709/
[2] http://docs.openstack.org/developer/ironic/dev/dev-quickstart.html

Change-Id: I5478e051372f6133cc902e82c4168774fb15b1df