v2 tokens validated on the v3 API are missing timezones
v3 tokens contain the issued_at and expires_at timestamps for each
token. If a token is created on the v2 API and then validated on the v3
API, this timezone information is missing (the 'Z' at the end of the
timestamp), and thus cannot be validated as ISO 8601 extended format
timestamps.  This appears to affect all token formats.

Change-Id: Ibf0e65852a010c595e9f1154695d3884549d2cbc
Closes-Bug: #1476329