Remove some translations
In preparation for the release, do a cleanup for translations.

Removes all po files that are partially translated. The translation
team has decided to exclude files with less than 66 % of translated
content. There is no content lost, all data is in the translation
server, we just remove it from the repository.

Statistics for removed files:

keystone/keystone-log-info.pot: Total strings 43
./keystone/locale/pt_BR/LC_MESSAGES/keystone-log-info.po: Translated: 19
./keystone/locale/pt_BR/LC_MESSAGES/keystone-log-info.po: Ratio 44
keystone/keystone-log-info.pot: Total strings 43
./keystone/locale/tr_TR/LC_MESSAGES/keystone-log-info.po: Translated: 23
./keystone/locale/tr_TR/LC_MESSAGES/keystone-log-info.po: Ratio 53

This change needs to be done manually since the automatic import does
not handle this.

Change-Id: I2e91afce06f52fe40fd77f41d535bf00a8eac302