Switch migration tests to oslo.db DbTestCase
oslo.db provides a DbTestCase base class that configures the oslo.db
for in-memory sqlite testing, in addition to being able to test
against other live databases.

Any support that we may have had for live database testing is
removed. It will be added back in a later patch to use features of
oslo.db DbTestCase to implement OpenStack's common way to do
live database testing.

Change-Id: I5b691d0adc812d687b318f639544ecaa165b9cfc