Correct `role_name` constraint dropping
The `role_name` constraint was not properly dropped in some cases
because the unique constraint was not consistently named. In all
cases we must search for the constraint expected, not assume the
name of the constraint will be consistent (especially from older
installs that have been moved forward in releases).

This change fixes migration 88, updates 96 (for backport to
stable/mitaka) This is being overly cautious, but specifically
is to handle the case where someone performed the upgrade and
manually fixed the migration resulting in duplicated constraints
in the role_table. Note that migration 100 was not backported since
it was a placeholder migration.

Conflicts:
    keystone/tests/unit/test_sql_upgrade.py

Co-Authored-By: "Matthew Thode" <mthode@mthode.org>
Change-Id: Ie0dc3d2449bace57d3e9323b281a2abd2ad0c983
closes-bug: #1562934
(cherry picked from commit 276e57e2083dcad8cbabf9aa9b3bd06c6079c415)