Random project should return positive numbers
The function random.randint(a, b) returns a random
number such that a <= N <= b .[0]

Since this test is returning random numbers such that
0 <= N <= len(projects)
It is possible to be subtracting 1 from 0 returning a random
number of -1.

Luckily Python is smart and allows the user to pass a negative number
as a index of a list to get the elements starting from  the right side :)
Which is technically not breaking anything and getting the same out
come, but we should probably do it correctly.

[0]: https://docs.python.org/2/library/random.html#random.randint

Change-Id: I63949aafd40f14d471f3fecf2895235254f36017