Fix totp test fails randomly
The test_with_multiple_users test would fail if the time happened to
roll over to the next chunk for the totp generator. The fix is to
control the clock in the test to this can't happen.

Change-Id: I2b92a0cc08ba8e36edc87cb76960a46746895458
(cherry picked from commit 3eaea2fdf417a03aeb539cd35ab28f01de5886af)