Make all fixture project_ids into uuids
The Fernet tests for Python 3.4 fail if they are given
project_ids that are not uuids. Since all issued
project IDs in a live deployment are UUIDs, it is more
correct to fix the tests than to change the formatter.

Change-Id: I485c02cbb6484e52b4bb4563e2842c45a34e66eb
(cherry picked from commit 36da34f02ff921584524108a34c11568bc406c10)