Fix confusing naming in ldap EnableEmuMixin.
In keystone/common/ldap/core.py, the get_all() method of EnabledEmuMixIn is
using "tenant" as the return object. This is misleading as the core method
is not using as project only but also other identity or resources (such as
user and group). We should change to name to be more generic to avoid
confusion.

Change-Id: I53a7eb22539697e16ac9893daf59506483345f9e
closes-bug: 1571342