Default caching to on for request-local caching.
Default cache to be enabled for request.local caching.
Since the default cache backend is dogpile.null, it is
safe to turn on caching by default. This will enable the
request_local caching for all deployments that do not
explicitly disable caching, thereby limiting the traffic
to the backends on a per-request basis.

With caching (in memcache) the addition of request local
cache was noticable (~10% inprovement), it stands to reason
that it will provide even more when it is offloading the
requests directly to SQL/LDAP.

Change-Id: I47d40c36799645d6fdd7b71e14e774ad75aa55f7