Drop the (unused) domain table
The Mitaka release removed all code that referenced the domain
table (with domains being stored instead as projects). However,
the physical domain table was not removed.

This patch removes the table. Such an action triggers our
detection of banned operations to ensure rolling upgrades are
possible. In this case, since all the code that referenced
this table was already removed in the previous release, this
is still a safe operation - so this migation is added to the
exception list in the test for banned operations.

Change-Id: I11bb23809e28b7fb0499cf79d261aa54f094b367