Remove helper script for py34
Remove the helper script for py34, it prevents the use of arguments
being passed down e.g. tox -epy34 -- --failing due to the lack of
passing {posargs} to the script. Since py34 has no special exceptions
now, the specific env for py34 with a special pretty_tox script is
not needed.

Change-Id: I4d960d82e2844c0c82a14be5e92230f1698e11b3