Isolate common ldap code to the identity backend
Since the LDAP code is now isolated to identity only move the common
LDAP code to be part of the ldap backend. This consolidates the code
to a specific location with a 2 cycle deprecation stub in
``keystone.common.ldap``. This lays the ground work for simplifying
the once extremely complex LDAP code that was originally written
to handle Identity and Assignment contructs.

Change-Id: I897bf24ba806c413bd17b76ed62a6cf446dfb0b6