Refactor revoke_model to remove circular dependency
In the process of caching service providers, a circular dependency was
uncovered between keystone's caching module and the revoke_model. This change
isolates the refactor to fix the circular dependency into it's own patch. A
separate follow on patch will add caching to service providers on token
creation and validation.

Change-Id: Iaa3c30d51270089b2d5999383d60ed0b2f7f56c8