Bootstrap: enable and reset password for existing users
One of the common use cases for the admin_token middleware was to
provide a recovery mechanism for cloud operators that had accidentally
disabled themselves or lost their password.

Instead of using bootstrap to create a second admin just to recover the
first, this change allows bootstrap to reset the user's credentials and
ensure that the account is enabled.

Change-Id: I82cafced67852335e9bb49035f13c993c7ccd2df
Closes-Bug: 1588860