Honor ldap_filter on filtered group list
Fix GET /v3/groups?name=<name> to honor conf.ldap.group_filter.

The case where groups are listed for a specific user was already
honoring the filter, but the case where all groups are listed was not.
Moved the check into the get_all_filtered method that is shared by both
cases so that it is not duplicated.

Change-Id: I4a11394de2e6414ba936e01bcf2fcc523bab8ba5
Closes-Bug: #1588927