Move TestAuth unscoped token tests to TokenAPITests
In test_v3_auth.py, we have a couple different classes that test similar token
API behaviors. The TestAuth class is only inherited and tested by the fernet
provider where the TokenAPITests class tested with the uuid, fernet, and pki
providers. This patch is one of many that will migrate duplicate tests from
TestAuth to TokenAPITests.

Change-Id: Ia2d7f857d078dde8486df6108009c76e50f353e0