Refactor test_validate_v2_scoped_token_with_v3_api
This commit moves the test_validate_v2_scoped_token_with_v3_api test from the
TestAuth module to the TokenAPITests module. It also renames the test to
accurately reflect what it is doing. Lastly, it removes a utility method,
get_v2_token() because it was no longer in use in TestAuth.

Change-Id: I95a19b2413740ac0c7e21636dd7f2f0dbbd00d69