Remove TestAuth
All token tests that describe the behavior of the token API have been
consolidated into the TokenAPITests class. This means that TestAuth is empty
and is no longer needed. This commit removes it and classes that inherited it
to run tests. There was only one class that inherited TestAuth and that was
TestAuthFernetTokenProvider. This class has also been removed since
TestFernetTokenAPIs exists and tests everything in the TokenAPITests.

All future tests that enhance or modify the behavior of token should be done
using TokenAPITests.

Change-Id: I1234aaffec6bea35e42e7d7367c397736498acfc