Verify domain_id when get_domain is being called
When create user using API it is possible to use a domain_id which
does match the created domain_id since mysql per default is not
case sensitive and returns the domain_id to be valid. In e.g.
liberty this breaks cli keystone v2 user list actions when a user
with a DEFaULt domain has been created.

With this change the domain_id is being validated with what provided
with the API call in get_domain.

cherry-picked from 7df92f7b624500e24b71c4b2d516604e0edb52f2

Change-Id: I028b2add3067e6fb9aa3f33eb8fe10d8ebace006
Closes-Bug: #1594284