Allow auth plugins to be setup more than once
I originally thought this was a good idea and long term I think it is.
The problem is that some tests need to use this fixture in the test
method itself and that will run after `load_backends()`, which also
tries to setup auth plugins.

This should come back after the tests are in good enough shapre to
support it.

Change-Id: I1175fa95d128940eb1a0feebdfe5d6e16285ad0a