Fixes failure when password is null
This patch allows the password column to be null which is the expected
behavior.

Note: We do have tests in place that should have caught this. However,
the SQL model for Password did not explicitly set the nullable column
attribute, so it defaulted to true, even though the migration set it to
false.

Closes-Bug: #1595819

Change-Id: I487072a5b1aeabebd1d2006781fa5dc568160ec5