Extracted common ldap setup and use in the filter tests
The common setup was taken from the BaseLDAPIdentity class as it holds
the setup for almost all of the LDAP tests. The goal is to make the
setup easier to follow. Currently it's hard to tell what happens during
setup...for instance, why is load_backends() called twice for some test
cases and three times for other?

Change-Id: I14dc4a2dc0d70fceef98bc5b6ab2103dddb54638