Adds a skip method to identify useless skips
We have places in our test code that skips tests in a child class
because the setup differs from the parent. This often leads to a
situation where we remove the original test, but since the skip
always passes we never remove it. This will make the skips fail if
the original method no longer exists.

Change-Id: I7f2049d21e1cbcbd3c74d85fd6c74680db2a245d