Rename [DEFAULT] keystone.conf module to keystone.conf.default
The renames the config module which default the [DEFAULT] configuration
section from 'base' (a name adopted based on nova's precedence with
nova.conf.base) to 'default', which should make it slightly easier for
developers to navigate the keystone.conf package, since all other
configuration modules directly reflect the name of the section they're
defining.

Also note, that we could take this one step further and call the module
keystone.conf.DEFAULT, but it's actually a PEP8 violation to name
module in uppercase (it reads like a reference to a constant in the
keystone.conf module). From PEP8:

    Modules should have short, all-lowercase names. Underscores can be
    used in the module name if it improves readability."

Source: https://www.python.org/dev/peps/pep-0008/#package-and-module-names

Change-Id: I78d4ecf67ddcb5c3c46ce0c85e1f8330fbf99877