Fixes hacking's handling of log hints
Our hacking check was being a little too specific when ensuring we are
using the proper hints. This code fails because we are expecting that
our code will not use a kwarg in the a raise case:

  msg = _('***')
  LOG.warning(msg)
  raise Exception(message=msg)

Hacking was forcing that msg used _LW(). This check accounts for kwargs
and allows the above code to work.

Change-Id: Id0e7a8dd2773e6c31e3ad034a46e414230312db9