Fix python{3,}-all-dev depends in deb based
This fixes inconsistencies in other-requirements.txt. For example,
python3-all-dev depends on python3-all, so we don't need to manually
install python3-all if we install python3-all-dev. python3-all means
all version of python3 available in the system, so we shouldn't
manually install python3.4 or 3.5. Also, using [platform:ubuntu-
is a bad idea (ie: it wont work in Debian).

Change-Id: Iaaaa2b137c1b751f8383ba21343dcc704c868187