Remove get_user_id in trust controller
The get_user_id function relied on the token model in auth_context -
which amongst other things means it would fail with a tokenless auth
context. This can be replaced with user_id from the request.context.

Normally I try not to mix in cleanups, but whilst doing this and
changing variables in list_trusts a small reshuffle and some whitespace
just made the whole thing actually readable.

Change-Id: I5b64210fc797961b422a0ab9a1b4cee078fe6a0f