v2 api: correct user list
The GET /v2.0/users route is unnecessarily in the OS-KSADM APIs, it's
not part of that extension. It is already defined correctly in the
v2-admin APIs.

Further, the sample response shows a single user, not a list, so fix
that while we're at it.

Route link: https://github.com/openstack/keystone/blob/master/keystone/v2_crud/admin_crud.py#L81-L86

Change-Id: I4131b955a6986a1311c1e100a54ab18f8622b2ff