v2 api: change update user
The PUT /v2.0/users/{userId} route is unnecessarily in the OS-KSADM APIs,
it's not part of that extension. It is already defined correctly in the
v2-admin APIs.

Further, the update example is the v2-ext docs was a bit nicer, so use
that one, update the request and response for that route. Delete the
example in v2-ext since nothing else uses it.

Route link: https://github.com/openstack/keystone/blob/master/keystone/v2_crud/admin_crud.py#L93-L97

Change-Id: I14bdf9a3b4c09a7b48cc52b0ff32f72dcb115178