Add dummy domain_id column to cached role
When token caching is turned on, upgrading from stable/liberty to
stable/mitaka or master causes tokens to fail to be issued for the
time-to-live of the cache. This is because as part of the token
issuance the token's role is looked up, and the cached version of the
role immediately after upgrade does not have a domain_id field, even
though that column was successfully added to the role database. This
patch hacks around that by artificially adding a null domain_id value
to the role reference.

This must be done in the manager, as opposed to the driver, because it
is the manager that is caching the value and so modifying the value
returned by the driver has no effect.

Change-Id: I55c791486f2a26ae995f693370b016895176a16f
Closes-bug: #1592169
(cherry picked from commit bc99dc76775d22eca01b818f37de35a76ece9d72)