Remove the extensions repos
All extension are now migrated into the keystone core and the
migrations in the old contrib folders just raise an error. Since
these can only be called from the now deprecated --extensions
options in keyston-manage, we might as well get rid of them all
and just raise the error in keystone manage. We can also clean
up the mirgation helpers accordingly

This clean up is useful ahead of layering on the rolling upgrade
support in keystone manage and the helpers.

Change-Id: I749dc7e13050e211df1f4fd144cb0479891e3e64