Remove mention of db_sync per backend
Many releases ago, we supported the notion of having migration
repositories per backend interface. If a backend needed to use the
database, it could manage it's own migrations independently from the
rest of keystone. That functionality was removed long ago, and this
blurb of text should have been removed with it.

Change-Id: If90e25ec556cf42322509ef28878e96120b0baad